Added UDP support

+++ socket_tutorial/socket_tutorial/client.h
@@ -30,5 +30,5 @@ void setMessageProperties(char messageProperties[], int bufferSize, int numberOf
 void fillMessageBuffer(char messageBuffer[], int bufferSize, int messageIndexOffset);
 double stopWatch(void);
 int calculateSpeed(int bytes, int elapsedTimeMS);
-void printResults(int size, int sent, int time, int speed, int missing, int disordered);
+void printResults(int size, int sent, int time, long unsigned int speed, int missing, int disordered);
 void convertDecToHex(int decimal, char* hexaNum);
